script for training using tpu

+++ workspace/train_model.sh
@@ -13,10 +13,14 @@ if [ -f .env ]; then
     source .env
 fi
 
+MODEL_DIR=models/${MODEL_NAME}/${MODEL_VERSION}
+PIPELINE_CONFIG_PATH=$MODEL_DIR/pipeline.config
+
 RESTART_CHECKPOINT=${RESTART_CHECKPOINT:-false}
 POSITION_ARGS=("")
 SYNC_TO_DRIVE=${SYNC_TO_DRIVE:-false}
 USE_TPU=${USE_TPU:-false}
+
 while [[ $# > 0 ]]; do
     case "$1" in
     --restart)
@@ -29,6 +33,7 @@ while [[ $# > 0 ]]; do
         ;;
     --tpu)
         USE_TPU=true
+        MODEL_DIR=$GS_PATH_PREFIX/$MODEL_DIR
         shift
         ;;
     *) # unknown flag/switch
@@ -40,16 +45,21 @@ done
 set -- "${POSITION_ARGS[@]}"
 
 if [ "$RESTART_CHECKPOINT" = "true" ]; then
-    cd models/${MODEL_NAME}/${MODEL_VERSION}
-    rm -rf checkpoint train ckpt-*
+    if [ "$USE_TPU" = "true" ]; then
+        # gsutil TODO
+        echo TODO
+    else
+        cd models/${MODEL_NAME}/${MODEL_VERSION}
+        rm -rf checkpoint train ckpt-*
+        cd $DIR
+    fi
 fi
-cd $DIR
 
-MODEL_DIR=models/${MODEL_NAME}/${MODEL_VERSION}
 
-SCRIPT_PATH=$DIR/../tensorflow_model_garden/research/object_detection/model_main_tf2.py
+# SCRIPT_PATH=$DIR/../tensorflow_model_garden/research/object_detection/model_main_tf2.py
+SCRIPT_PATH=$DIR/model_main_tf2.py
 python ${SCRIPT_PATH} \
-  --pipeline_config_path=models/${MODEL_NAME}/${MODEL_VERSION}/pipeline.config\
+  --pipeline_config_path=${PIPELINE_CONFIG_PATH} \
   --model_dir=${MODEL_DIR}\
   --checkpoint_every_n=100 \
   --num_workers=3 \
